Moved ItemNotFoundException to main package because its used for requestDB
