+++ src/main/java/edu/wpi/teamR/pathfinding/Pathfinder.java
@@ -1,6 +1,6 @@
 package edu.wpi.teamR.pathfinding;
 
-import edu.wpi.teamR.mapdb.ItemNotFoundException;
+import edu.wpi.teamR.ItemNotFoundException;
 import edu.wpi.teamR.mapdb.MapDatabase;
 import edu.wpi.teamR.mapdb.Node;
